Fix clang-tidy theme 6: modernize C++ syntax
- modernize-return-braced-init-list (35): Type() → {} in returns
- modernize-concat-nested-namespaces (9): C++17 A::B::C syntax
- modernize-use-nullptr (9): NULL/0 → nullptr in pointer contexts
- modernize-use-using (6): typedef → using alias
- modernize-use-override (2): add override to virtual destructors
- modernize-use-std-numbers (2): std::numbers::pi_v and sqrt2
- modernize-use-auto (1): auto with cast initializer
- modernize-use-integer-sign-comparison (1): size_t → int loop var

+++ src/include/TempLat/lattice/field/collections/vectorfieldcollection.h
@@ -23,7 +23,7 @@ namespace TempLat
                           LatticeParameters<T> pLatPar = LatticeParameters<T>()) //:
     {
       if constexpr (N > 0)
-        for (size_t i = 0; i < N; ++i) {
+        for (int i = 0; i < N; ++i) {
           fs.push_back(VectorField<Arg, T, NDim, flatAssign>(name + "_" + std::to_string(i), toolBox, pLatPar));
         }
     }
